+++ CHANGELOG.md
@@ -9,7 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `Mastermind` Archetype (Isekai Protagonist).
 - Added `Overlord` Archetype (Isekai Protagonist).
 ### Changed
-- Changed `Edge Lord` to get another OP ability at 10th level.
+- Buffed `Isekai Protagonist` to be able to choose between `Dark Aura` and `Friendly Aura`.
+- Buffed `Edge Lord` to get another OP ability at 10th level.
+- Buffed `God Emperor` to get another OP ability at 5th and 15th level.
+- Buffed `Hero` to get another Special Power at 3rd and 17th level.
+- Buffed `Hero` to get Tower shield proficiency.
 ### Removed
 - Removed `Villain` Archetype (Isekai Protagonist).
 - Removed `Kinetic Power` Special Power.